+++ src/assets/scripts/client/aircraft/AircraftCommander.js
@@ -229,11 +229,11 @@ export default class AircraftCommander {
         const airport = AirportController.airport_get();
 
         return aircraft.pilot.maintainAltitude(
-            aircraft.altitude,
             altitudeRequested,
             expediteRequested,
             shouldUseSoftCeiling,
-            airport
+            airport,
+            aircraft
         );
     }
 
@@ -308,9 +308,9 @@ export default class AircraftCommander {
      * @param data {array}
      */
     runSpeed(aircraft, data) {
-        const speed = data[0];
+        const nextSpeed = data[0];
 
-        return aircraft.pilot.maintainSpeed(aircraft.speed, speed);
+        return aircraft.pilot.maintainSpeed(nextSpeed, aircraft);
     }
 
     /**